Feat/apply dynamic metadata (#22)
# Overview
결과에 따라 다른 metadata를 전송합니다.
